Add files via upload
This is an updated program for GEIS measurement during charging and discharging. The change was made for measure the potential of an electrochemical cell or battery. This program does not use Ei.Potential function, because I found it does not work correctly, or it is not updating during measurement. Instead of Ei.Potential, the program calls OCP measurement procedure with *.nox extension created in the NOVA program. The program uses EcoChemie SDK dll libraries and WinUsbNet library from the NuGet Package manager. I am only not sure for Ei.SetPoint function, which changes DC electric current in galvanostatic mode from charging to discharging.