Add files via upload
This GEIS (galvanostatic electrochemical impedance spectroscopy) program uses Autolab from Metroohm. It does GEIS in charging. If the OCP or voltage on the battery exceeds the maximum temperature, it enters discharging mode and measures GEIS during discharging. The program before each EIS measures OCP and writes it to the voltage_data text file. The time stamp is saved to the time_data text file at the beginning of each EIS measurement.